- Diff Img: in config_preprocessing.yaml, added field `exclude_neighbor_objs_outside` and removed fields `qual_metrics_tbl_fp` and `sat_tbl_fp`; set `upscale_f` to from (3, 3) to (5, 5) to increase final resolution to 55x55 px. - Diff Img: moved functions used for preprocessing of difference image data from preprocess_diff_img.py to utils_diff_img.py. - Diff Img: in preprocess_diff_img.py, added neighbor_data as argument to `preprocess_single_diff_img_data_for_example()`; target and neighbors coordinate transformation is done outside the functions used for each preprocessing step; added step for creating neighbors image. - Diff Img: in preprocess_diff_img.py, updated `preprocess_diff_img_tces()`. This involved changing the way the sampling of the quarters/sector is conducted. Before, you could have the same preprocessing happening for the same quarter/sector multiple times in the case that quarter/sector was sampled more than one. Now the processing happens only for the unique quarters/sectors that were sampled; change plot function to show neighbors image and be more closely related to the plot obtained from the extraction step; changed columns in the auxiliary table. - Diff Img: in search_neighbors_and_map_location_to_ccd.py, added option to map neighbors coordinates to the CCD frame using the WCS transformation stored in the target pixel file of the target star; added option in `get_ccd_coords_neighbors_targets()` to use existing targets NumPy file if it aleady exists; same thing for the target_sectors_pairs.csv file. - Diff Img: in utils_diff_img.py, changed extraction of quality metric information to convert string values in the DV xml file to float and boolean variables

+++ src_preprocessing/diff_img/search_neighbors_and_map_location_to_ccd.py
@@ -12,7 +12,8 @@ import pandas as pd
 # import subprocess
 import multiprocessing
 import logging
-from tess_stars2px import tess_stars2px_function_entry
+# from tess_stars2px import tess_stars2px_function_entry
+import lightkurve as lk
 
 
 def log_or_print(message, logger=None):
@@ -91,7 +92,7 @@ def filter_neighbors_by_magnitude(neighbors_tbl, mag_thr):
 
 
 def create_targets_to_neighbors_data(neighbors_tbl, save_fp):
-    """ Create dictionary that maps targets to their neighbors. The dictionary maps target ID to a subdictionary with
+    """ Create dictionary that maps targets to their neighbors. The dictionary maps target ID to a sub-dictionary with
     their neighbors. Each neighbor ID maps to a dictionary containing 'dstArcSec', the distance between the two starts
     in arcseconds, and 'mag`, the magnitude of the neighboring star.
 
@@ -113,7 +114,7 @@ def create_targets_to_neighbors_data(neighbors_tbl, save_fp):
     np.save(save_fp, targets_dict)
 
 
-def map_star_celestial_coordinates_to_ccd(stars_tbl, sector=None):
+def map_star_celestial_coordinates_to_ccd(stars_tbl, sector=None, logger=None):
     """ Maps celestial coordinates of stars (RA and Dec) in degrees to the pixel coordinates for each star in
     one/multiple observed sectors (columns 'col_px' and 'row_px').
 
@@ -121,23 +122,44 @@ def map_star_celestial_coordinates_to_ccd(stars_tbl, sector=None):
         stars_tbl: pandas DataFrame, table with stars. Must contain columns 'ID', 'ra', and 'dec', with RA and
         Dec in degrees
         sector: int, sector for which to get the pointing. If `None`, it considers all available sectors
+        logger: logger
 
     Returns:
         stars_px_coords_tbl: pandas DataFrame, CCD pixel coordinates for each star in one/multiple observed sectors
 
     """
 
-    import timeit
+    # using WCS transformation stored in the target pixel files for each object
+    outColPix, outRowPix, stars_ids = [], [], stars_tbl['ID']
+    for target_id, objs in stars_tbl.groupby('target_id'):
 
-    # def _wrapper_fn():
+        search_result = lk.search_targetpixelfile(target=f'tic{target_id}', mission='TESS',
+                                                  author=('TESS-SPOC', 'SPOC'),
+                                                  exptime=120, sector=sector)
 
-    # input_fp = save_fp.parent / f'{save_fp.stem}_input_to_tess_point.txt'
+        if len(search_result) > 1:
+            log_or_print(f'Found {len(search_result)} target pixel files for target {target_id} in sector '
+                         f'{sector}', logger)
 
-    # initialize scinfo
-    _, _, _, _, _, _, _, _, scinfo = tess_stars2px_function_entry(stars_tbl['ID'][0], stars_tbl['ra'][0], stars_tbl['dec'][0], trySector=sector)
+        # Download the first search result
+        tpf = search_result.download(download_dir='/home/msaragoc/Downloads/', quality_bitmask='default')
+        # use WCS transformation to map objects to CCD pixel frame
+        outColPix_obj, outRowPix_obj = tpf.wcs.all_world2pix(objs['ra'], objs['dec'], 0)
 
-    stars_ids, _, _, _, _, _, outColPix, outRowPix, scinfo = (
-        tess_stars2px_function_entry(stars_tbl['ID'], stars_tbl['ra'], stars_tbl['dec'], scInfo=scinfo, trySector=sector))
+        outColPix += list(outColPix_obj)
+        outRowPix += list(outRowPix_obj)
+
+        # delete target pixel file
+        Path(tpf.path).unlink()
+
+    # # using tess-point
+    # # initialize scinfo
+    # _, _, _, _, _, _, _, _, scinfo = tess_stars2px_function_entry(stars_tbl['ID'][0], stars_tbl['ra'][0],
+    #                                                               stars_tbl['dec'][0], trySector=sector)
+    #
+    # stars_ids, _, _, _, _, _, outColPix, outRowPix, scinfo = (
+    #     tess_stars2px_function_entry(stars_tbl['ID'], stars_tbl['ra'], stars_tbl['dec'], scInfo=scinfo,
+    #                                  trySector=sector))
 
     stars_px_coords_tbl = pd.DataFrame(
         {'ID': stars_ids,
@@ -165,13 +187,12 @@ def create_neighbors_table(neighbors_tbl, save_fp, sector=None, logger=None):
     """
 
     neighbors = neighbors_tbl.drop_duplicates('ID')  # drop duplicates
-    neighbors = neighbors.drop(columns=['dstArcSec', 'target_id'])
+    neighbors = neighbors.drop(columns=['dstArcSec'])  # , 'target_id'])
     log_or_print(f'Found {len(neighbors)} unique neighbors.', logger)
 
     # map celestial coordinates to CCD from tess-point for these neighbors
-    # create txt file as input to tess-point
-    log_or_print('Getting pixel coordinates for neighbors using tess-point...', logger)
-    stars_px_coords_tbl = map_star_celestial_coordinates_to_ccd(neighbors, sector)
+    log_or_print('Getting pixel coordinates for neighbors...', logger)
+    stars_px_coords_tbl = map_star_celestial_coordinates_to_ccd(neighbors, sector, logger)
 
     # add pixel coordinates to neighbors table
     log_or_print(f'Adding pixel coordinates to neighbors...', logger)
@@ -214,17 +235,29 @@ def get_ccd_coords_neighbors_targets(targets, search_radius_arcsec, mag_thr, res
 
     log_or_print(f'Searching for neighbors in a radius of {search_radius_arcsec} arcsec '
                  f'for {len(targets)} targets...', logger)
-    search_res_targets = get_neighbors_in_search_radius(targets, search_radius_arcsec, mag_thr, logger)
 
-    # log_or_print(f'Filtering neighbors based on magnitude: magnitude threshold = {mag_thr}...', logger)
-    # search_res_tics_mag = filter_neighbors_by_magnitude(search_res_targets, mag_thr)
+    if (res_dir / f'targets_S{res_suffix}.npy').exists():
+        log_or_print(f'Found NumPy file '
+                     f'for targets in sector {res_suffix}: {res_dir / f"targets_S{res_suffix}.npy"}. Loading data...',
+                     logger)
+        targets_dict = np.load(res_dir / f'targets_S{res_suffix}.npy', allow_pickle=True).item()
+        search_res_targets = pd.DataFrame.from_dict({(i, j): targets_dict[i][j]
+                                                     for i in targets_dict.keys()
+                                                     for j in targets_dict[i].keys()},
+                                                    orient='index')
+        search_res_targets.index.names = ['target', 'ID']
+        search_res_targets = search_res_targets.reset_index()
 
-    # save_fp_targets = res_dir / f'targets{tbl_suffix}_search_radius_{search_radius_arcsec.value}_mag_thr_{mag_thr}.npy'
-    log_or_print('Creating target data based on search results...', logger)
-    save_fp_targets = res_dir / f'targets_S{res_suffix}.npy'
-    create_targets_to_neighbors_data(search_res_targets, save_fp_targets)
+    else:
+        search_res_targets = get_neighbors_in_search_radius(targets, search_radius_arcsec, mag_thr, logger)
 
-    # save_fp_neighbors = res_dir / f'neighbors{tbl_suffix}_search_radius_{search_radius_arcsec.value}_mag_thr_{mag_thr}.csv'
+        # log_or_print(f'Filtering neighbors based on magnitude: magnitude threshold = {mag_thr}...', logger)
+        # search_res_tics_mag = filter_neighbors_by_magnitude(search_res_targets, mag_thr)
+
+        log_or_print('Creating target data based on search results...', logger)
+        save_fp_targets = res_dir / f'targets_S{res_suffix}.npy'
+        create_targets_to_neighbors_data(search_res_targets, save_fp_targets)
+
     log_or_print('Creating table with neighbors...', logger)
     save_fp_neighbors = res_dir / f'neighbors_S{res_suffix}.csv'
     create_neighbors_table(search_res_targets, save_fp_neighbors, int(res_suffix), logger)
@@ -242,11 +275,12 @@ if __name__ == "__main__":
     ]
     tce_tbl = pd.read_csv('/data5/tess_project/Data/Ephemeris_tables/TESS/DV_SPOC_mat_files/preprocessing_tce_tables/09-25-2023_1608/tess_2min_tces_dv_s1-s68_all_msectors_11-29-2023_2157_newlabels_nebs_npcs_bds_ebsntps_to_unks_sg1master_allephemmatches_exofoptois.csv',
                           usecols=tce_tbl_cols)
+    # tce_tbl = tce_tbl.loc[tce_tbl['target_id'] == 356229935]
     # search parameters
     search_radius_arcsec = 168 * u.arcsec  # np.sqrt(2) * 121 / 2 * u.arcsec  # 168 DV (Joe)
     mag_thr = np.inf
     # results directory
-    res_dir = Path(f'/home/msaragoc/Projects/exoplnt_dl/experiments/search_neighboring_stars/tess_spoc_2min_s1-s68_search_radius_arcsec_{search_radius_arcsec.value}_mag_thr_{mag_thr}_1-30-2025_1212')
+    res_dir = Path(f'/home/msaragoc/Projects/exoplnt_dl/experiments/search_neighboring_stars/tess_spoc_2min_s1-s68_search_radius_arcsec_{search_radius_arcsec.value}_tpf_wcs_2-10-2025_1008')
     # multiprocessing parameters
     # n_jobs = 1
     n_procs = 13
@@ -256,32 +290,38 @@ if __name__ == "__main__":
     res_dir.mkdir(parents=True, exist_ok=True)
 
     # get sectors observed for each target in the table
-    targets_dict = {'target_id': [], 'sector': [], 'mag': []}
-    targets = tce_tbl.groupby('target_id')
-    for target_id, target_data in targets:
-        target_data = target_data.reset_index(drop=True)
-        obs_sectors_target = np.nonzero(np.prod(
-            target_data['sectors_observed'].apply(lambda x: np.array([int(el) for el in x])).values, axis=0))[0]
-        for obs_sector in obs_sectors_target:
-            targets_dict['target_id'].append(target_id)
-            targets_dict['sector'].append(obs_sector)
-            targets_dict['mag'].append(target_data.loc[0, 'mag'])
+    if (res_dir / 'target_sector_pairs.csv').exists():
+        print('Reading table for target-sector pairs already created.')
+        targets_df = pd.read_csv(res_dir / 'target_sector_pairs.csv')
+    else:
+        targets_dict = {'target_id': [], 'sector': [], 'mag': []}
+        targets = tce_tbl.groupby('target_id')
+        for target_id, target_data in targets:
+            target_data = target_data.reset_index(drop=True)
+            obs_sectors_target = np.nonzero(np.prod(
+                target_data['sectors_observed'].apply(lambda x: np.array([int(el) for el in x])).values, axis=0))[0]
+            for obs_sector in obs_sectors_target:
+                targets_dict['target_id'].append(target_id)
+                targets_dict['sector'].append(obs_sector)
+                targets_dict['mag'].append(target_data.loc[0, 'mag'])
 
-    targets_df = pd.DataFrame(targets_dict)
-    targets_df.to_csv(res_dir / 'target_sector_pairs.csv', index=False)
+        targets_df = pd.DataFrame(targets_dict)
+        targets_df.to_csv(res_dir / 'target_sector_pairs.csv', index=False)
+
     print(f'Found {len(targets_df)} target-sector pairs.')
 
     # targets_df = targets_df.loc[targets_df['sector'] == 2][:3]
 
     # split based on sector runs
     tics_jobs = {sector: targets_srun.reset_index(drop=True)
-                 for sector, targets_srun in targets_df.groupby('sector')}
+                 for sector, targets_srun in targets_df.groupby('sector')}  #  if sector in [12, 38, 3, 7, 9]}
     print(f'Found {len(tics_jobs)} sectors.')
 
     n_jobs = len(tics_jobs)
     print(f'Split work into {n_jobs} jobs.')
 
     # parallelize jobs
+    n_procs = min(n_jobs, n_procs)
     pool = multiprocessing.Pool(processes=n_procs)
     jobs = [(tics_job_sector_run, search_radius_arcsec, mag_thr, res_dir, str(sector_run), use_logs)
             for sector_run, tics_job_sector_run in tics_jobs.items()]
